Move FolderMode out of the account object
