+++ legacy/account/src/main/java/app/k9mail/legacy/account/Account.kt
@@ -601,14 +601,6 @@ open class Account(
         return uuid.hashCode()
     }
 
-    enum class FolderMode {
-        NONE,
-        ALL,
-        FIRST_CLASS,
-        FIRST_AND_SECOND_CLASS,
-        NOT_SECOND_CLASS,
-    }
-
     enum class SpecialFolderSelection {
         AUTOMATIC,
         MANUAL,
