+++ legacy/account/src/main/java/app/k9mail/legacy/account/FolderMode.kt
@@ -1,0 +1,9 @@
+package app.k9mail.legacy.account
+
+enum class FolderMode {
+    NONE,
+    ALL,
+    FIRST_CLASS,
+    FIRST_AND_SECOND_CLASS,
+    NOT_SECOND_CLASS,
+}
