+++ legacy/account/src/test/kotlin/app/k9mail/legacy/account/LegacyAccountWrapperTest.kt
@@ -2,7 +2,6 @@ package app.k9mail.legacy.account
 
 import app.k9mail.legacy.account.Account.DeletePolicy
 import app.k9mail.legacy.account.Account.Expunge
-import app.k9mail.legacy.account.Account.FolderMode
 import app.k9mail.legacy.account.Account.MessageFormat
 import app.k9mail.legacy.account.Account.QuoteStyle
 import app.k9mail.legacy.account.Account.ShowPictures
